refactor: rename comment to comments
