+++ src/posts/comments/comments.controller.ts
@@ -20,7 +20,7 @@ import {
 import { ReportsService } from 'src/reports/reports.service';
 import { RecommendsService } from 'src/recommends/recommends.service';
 import { RequestUser } from 'src/global/decorator/request-user.decorator';
-import { CommentService } from './comment.service';
+import { CommentService } from './comments.service';
 import { CreateCommentDto } from './dto/create-comment.dto';
 import { UpdateCommentDto } from './dto/update-comment.dto';
 import { ReportCommentDto } from './dto/report-comment.dto';
