+++ src/posts/comments/comments.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { CommentService } from './comment.service';
+import { CommentService } from './comments.service';
 import { AuthModule } from 'src/auth/auth.module';
 import { PrismaService } from 'src/prisma.service';
