+++ src/posts/posts.module.ts
@@ -5,8 +5,8 @@ import { PrismaService } from 'src/prisma.service';
 import { AuthModule } from 'src/auth/auth.module';
 import { ReportsModule } from 'src/reports/reports.module';
 import { RecommendsModule } from 'src/recommends/recommends.module';
-import { CommentModule } from './comment/comment.module';
-import { CommentsController } from './comment/comment.controller';
+import { CommentsController } from './comments/comments.controller';
+import { CommentModule } from './comments/comments.module';
 
 @Module({
   imports: [AuthModule, RecommendsModule, ReportsModule, CommentModule],
